Extract Command+Utilities + cleanup
