+++ Sources/VaporToolbox/Command+Utilities.swift
@@ -1,16 +1,8 @@
+
 import Foundation
 import Console
 
-internal let defaultTemplatesDirectory = ".build/Templates/"
-internal let defaultTemplatesURLString = "https://gist.github.com/1b9b58c0ca4dbe3538b2707df5959d80.git"
-
-public protocol Generator {
-    var console: ConsoleProtocol { get }
-    init(console: ConsoleProtocol)
-    func generate(arguments: [String]) throws
-}
-
-public extension Generator {
+public extension Command {
 
     public func fileExists(atPath path: String) -> Bool {
         return FileManager.default.fileExists(atPath: path)
